renommed old CheckersAI into AlphaBetaAI and moved into fr.uparis.checkers_ai.ai package, created new CheckersAI
